Update Icepap example
Creating a custom class is no longer needed.
